// src/redux/filter/types.ts
export enum SortPropertyEnum {
	RATING_DESC = "rating",
	RATING_ASC = "-rating",
	PRICE_DESC = "price",
	PRICE_ASC = "-price",
	TITLE_DESC = "title",
	TITLE_ASC = "-title",
}

export type Sort = {
	name: string
	sort: SortPropertyEnum
}

export interface FilterSliceState {
	searchValue: string
	categoryType: number
	activePage: number
	sortType: Sort
}